// src/types/portfolio.ts
export type Link = {
  href: string;
  label: string;
};

export type Highlight = {
  id: string;
  value: string;
  label: string;
  description: string;
};

export type Service = {
  id: string;
  title: string;
  description: string;
};

export type Certificate = {
  id: string;
  image: string;
  title: string;
  description: string;
};

export type Skill = {
  id: string;
  image: string;
  label: string;
};